MAJOR CHANGES
I am removing a lot of code/classes/interfaces and reorganizing the rest in an effort to the reduce the size of the lib when it's used in projects.

If you want to take advantage of these changes then you should use the new YamlDecoder.as class, instead of YAML.decode();  Right now I've only updated the decoder, I am still working on the encoder.  If you use the YamlDecoder.as class then only the classes necessary for decoding will be included in your project.  The end result is that if you are just using the decoder then the yaml lib will only increase the size of your project about 25k, rather than the 70+k it used to add.

I've also remove all references to the Flex lib, so this code should now work out of the box with Flash and straight AS3 projects.